requirements.txt
asgiref==3.3.1
Django==3.1.5
django-appconf==1.0.4
pilkit==2.0
Pillow==8.1.0
pytz==2020.5
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
